refactor : Test 리펙토링
