+++ src/main/java/chess/domain/chessPiece/piece/Bishop.java
@@ -1,22 +1,22 @@
-package chess.domain.chessPiece;
+package chess.domain.chessPiece.piece;
 
 import chess.domain.Position;
 import chess.domain.chessPiece.team.TeamStrategy;
+import chess.domain.movefactory.CrossType;
 import chess.domain.movefactory.MoveType;
 
-public class King extends Piece {
-
-    public King(Position position, TeamStrategy teamStrategy) {
+public class Bishop extends Piece {
+    public Bishop(Position position, TeamStrategy teamStrategy) {
         super(position, teamStrategy);
     }
 
     @Override
     public boolean isMovable(MoveType moveType) {
-        return true;
+        return moveType instanceof CrossType;
     }
 
     @Override
     public String pieceName() {
-        return teamStrategy.kingName();
+        return teamStrategy.bishopName();
     }
 }
