+++ src/main/java/chess/domain/chessPiece/piece/Knight.java
@@ -1,22 +1,23 @@
-package chess.domain.chessPiece;
+package chess.domain.chessPiece.piece;
 
 import chess.domain.Position;
 import chess.domain.chessPiece.team.TeamStrategy;
-import chess.domain.movefactory.CrossType;
 import chess.domain.movefactory.MoveType;
 
-public class Bishop extends Piece {
-    public Bishop(Position position, TeamStrategy teamStrategy) {
+public class Knight extends Piece {
+
+    public Knight(Position position, TeamStrategy teamStrategy) {
         super(position, teamStrategy);
     }
 
     @Override
     public boolean isMovable(MoveType moveType) {
-        return moveType instanceof CrossType;
+        return true;
+        //TODO 미구현 ;
     }
 
     @Override
     public String pieceName() {
-        return teamStrategy.bishopName();
+        return teamStrategy.knightName();
     }
 }
