+++ src/main/java/chess/domain/chessPiece/piece/Piece.java
@@ -1,4 +1,4 @@
-package chess.domain.chessPiece;
+package chess.domain.chessPiece.piece;
 
 import chess.domain.Position;
 import chess.domain.chessPiece.team.TeamStrategy;
@@ -24,15 +24,6 @@ public abstract class Piece implements PieceAbility {
 
     public boolean isSameTeam(Piece targetPiece) {
         return teamStrategy.equals(targetPiece.teamStrategy);
-    }
-
-    protected void validSameTeam(Piece targetPiece) {
-        if (targetPiece == null) {
-            return;
-        }
-        if (isSameTeam(targetPiece)) {
-            throw new IllegalArgumentException("같은팀이 있는 칸으로 이동할수 없습니다.");
-        }
     }
 }
