+++ src/main/java/chess/domain/chessPiece/piece/Queen.java
@@ -1,23 +1,22 @@
-package chess.domain.chessPiece;
+package chess.domain.chessPiece.piece;
 
 import chess.domain.Position;
 import chess.domain.chessPiece.team.TeamStrategy;
 import chess.domain.movefactory.MoveType;
-import chess.domain.movefactory.StraightType;
 
-public class Rook extends Piece {
-    public Rook(Position position, TeamStrategy teamStrategy) {
+public class Queen extends Piece {
+
+    public Queen(Position position, TeamStrategy teamStrategy) {
         super(position, teamStrategy);
     }
 
-
     @Override
     public boolean isMovable(MoveType moveType) {
-        return moveType instanceof StraightType;
+        return true;
     }
 
     @Override
     public String pieceName() {
-        return teamStrategy.rookName();
+        return teamStrategy.queenName();
     }
 }
