+++ src/main/java/chess/domain/chessPiece/piece/Rook.java
@@ -1,22 +1,23 @@
-package chess.domain.chessPiece;
+package chess.domain.chessPiece.piece;
 
 import chess.domain.Position;
 import chess.domain.chessPiece.team.TeamStrategy;
 import chess.domain.movefactory.MoveType;
+import chess.domain.movefactory.StraightType;
 
-public class Queen extends Piece {
-
-    public Queen(Position position, TeamStrategy teamStrategy) {
+public class Rook extends Piece {
+    public Rook(Position position, TeamStrategy teamStrategy) {
         super(position, teamStrategy);
     }
 
+
     @Override
     public boolean isMovable(MoveType moveType) {
-        return true;
+        return moveType instanceof StraightType;
     }
 
     @Override
     public String pieceName() {
-        return teamStrategy.queenName();
+        return teamStrategy.rookName();
     }
 }
