+++ src/test/java/chess/domain/chessPiece/piece/KnightTest.java
@@ -1,0 +1,24 @@
+package chess.domain.chessPiece.piece;
+
+class KnightTest {
+//    @Test
+//    @DisplayName("이동 성공 테스트")
+//    void movable() {
+//        Position position = Position.of(File.D, Rank.TWO);
+//        Position target = Position.of(File.B, Rank.ONE);
+//        Piece knight = new Knight(position, new BlackTeam());
+//
+//        assertThat(knight.isMovable(target, null)).isTrue();
+//    }
+//
+//    @Test
+//    @DisplayName("이동 실패 테스트")
+//    void isNotMovable() {
+//        Position position = Position.of(File.D, Rank.TWO);
+//        Position target = Position.of(File.C, Rank.ONE);
+//        Piece knight = new Knight(position, new BlackTeam());
+//
+//        assertThat(knight.isMovable(target, null)).isFalse();
+//    }
+
+}
