+++ src/test/java/chess/domain/chessPiece/piece/PawnTest.java
@@ -1,0 +1,35 @@
+package chess.domain.chessPiece.piece;
+
+class PawnTest {
+
+//    @Test
+//    @DisplayName("폰 처음 2칸이동 테스트")
+//    void isMovable() {
+//        ChessBoard chessBoard = new ChessBoard();
+//        Position sourcePosition = Position.of("b2");
+//        Position targetPosition = Position.of("b4");
+//
+//        Piece blackPawn = chessBoard.findPieceByPosition(sourcePosition);
+//        MoveType moveType = MoveFactory.of(sourcePosition, targetPosition);
+//
+//        assertThat(blackPawn.isMovable(moveType)).isTrue();
+//    }
+//
+//    @Test
+//    @DisplayName("폰 공격 테스트")
+//    void isMovableAttack() {
+//        Piece blackPawn = new Pawn(Position.of("b2"), new BlackTeam());
+//        Piece whitePawn = new Pawn(Position.of("c3"), new WhiteTeam());
+//
+//        assertThat(blackPawn.isMovable(Position.of("c3"), whitePawn)).isTrue();
+//    }
+//
+//    @Test
+//    @DisplayName("폰 뒤로 이동 테스트")
+//    void isMovableFalse() {
+//        Piece blackPawn = new Pawn(Position.of("b2"), new BlackTeam());
+//
+//        assertThat(blackPawn.isMovable(Position.of("b1"), null)).isFalse();
+//
+//    }
+}
